Refactor code by ViewAdapter but i take some error at view grid, largeGrid and Detail Adapter
