+++ app/src/main/java/com/example/colornote/adapter/ViewDetailsAdapter.java
@@ -5,6 +5,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
+import android.widget.GridView;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -24,15 +25,11 @@ import com.example.colornote.viewpager.CustomViewEmpty;
 
 import java.util.ArrayList;
 
-public class ViewDetailsAdapter extends BaseAdapter {
-    ArrayList<Task> tasks;
-    Context context;
-    ColorDAO colorDAO;
+public class ViewDetailsAdapter extends ViewAdapter {
 
     public ViewDetailsAdapter(ArrayList<Task> tasks, Context context){
-        this.tasks = tasks;
-        this.context = context;
-        colorDAO = ColorDAO.getInstance();
+        super(tasks, context);
+
     }
 
     @Override
@@ -52,6 +49,7 @@ public class ViewDetailsAdapter extends BaseAdapter {
 
     @Override
     public View getView(int position, View  view, ViewGroup parent) {
+        this.parent = (GridView) parent;
         SelectedObserverService.getInstance().setSelected(new boolean[tasks.size()]);
         ViewHolder holder = null;
         Task task = tasks.get(position);
@@ -86,12 +84,8 @@ public class ViewDetailsAdapter extends BaseAdapter {
             @Override
             public boolean onLongClick(View v) {
                 if(SelectedObserverService.getInstance().getIsSelected()[position] == false){
-                    ((CustomCardView) v).addBorder();
-                    ((CustomViewEmpty) finalHolder.colorSub).addBorder();
                     SelectedObserverService.getInstance().selected(position, position+1);
                 }else{
-                    ((CustomCardView) v).removeBorder();
-                    ((CustomViewEmpty) finalHolder.colorSub).removeBorder();
                     SelectedObserverService.getInstance().unselected(position, position+1);
                 }
 
@@ -104,14 +98,11 @@ public class ViewDetailsAdapter extends BaseAdapter {
             public void onClick(View v) {
                 if(SelectedObserverService.getInstance().hasSelected()){
                     if(SelectedObserverService.getInstance().getIsSelected()[position] == false){
-                        ((CustomCardView) v).addBorder();
-                        ((CustomViewEmpty) finalHolder.colorSub).addBorder();
                         SelectedObserverService.getInstance().selected(position, position+1);
                     }else{
-                        ((CustomCardView) v).removeBorder();
-                        ((CustomViewEmpty) finalHolder.colorSub).removeBorder();
                         SelectedObserverService.getInstance().unselected(position, position+1);
                     }
+                    updateBorderView();
                 }
             }
         });
