+++ app/src/main/java/com/example/colornote/adapter/ViewLargeGridAdapter.java
@@ -5,6 +5,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
+import android.widget.GridView;
 import android.widget.ImageView;
 import android.widget.TextView;
 
@@ -23,15 +24,10 @@ import com.example.colornote.viewpager.CustomViewEmpty;
 
 import java.util.ArrayList;
 
-public class ViewLargeGridAdapter extends BaseAdapter {
-    ArrayList<Task> tasks;
-    Context context;
-    ColorDAO colorDAO;
+public class ViewLargeGridAdapter extends ViewAdapter {
 
     public ViewLargeGridAdapter(ArrayList<Task> tasks, Context context){
-        this.tasks = tasks;
-        this.context = context;
-        colorDAO = ColorDAO.getInstance();
+        super(tasks, context);
     }
 
     @Override
@@ -51,6 +47,7 @@ public class ViewLargeGridAdapter extends BaseAdapter {
 
     @Override
     public View getView(int position, View  view, ViewGroup parent) {
+        this.parent = (GridView) parent;
         SelectedObserverService.getInstance().setSelected(new boolean[tasks.size()]);
         ViewHolder holder = null;
         Task task = tasks.get(position);
@@ -81,14 +78,11 @@ public class ViewLargeGridAdapter extends BaseAdapter {
             @Override
             public boolean onLongClick(View v) {
                 if(SelectedObserverService.getInstance().getIsSelected()[position] == false){
-                    ((CustomCardView) v).addBorder();
-                    ((CustomViewEmpty) finalHolder.colorSub).addBorder();
                     SelectedObserverService.getInstance().getIsSelected()[position] = true;
                 }else{
-                    ((CustomCardView) v).removeBorder();
-                    ((CustomViewEmpty) finalHolder.colorSub).removeBorder();
                     SelectedObserverService.getInstance().getIsSelected()[position] = false;
                 }
+                updateBorderView();
 
                 return true;
             }
@@ -99,14 +93,11 @@ public class ViewLargeGridAdapter extends BaseAdapter {
             public void onClick(View v) {
                 if(SelectedObserverService.getInstance().hasSelected()){
                     if(SelectedObserverService.getInstance().getIsSelected()[position] == false){
-                        ((CustomCardView) v).addBorder();
-                        ((CustomViewEmpty) finalHolder.colorSub).addBorder();
                         SelectedObserverService.getInstance().getIsSelected()[position] = true;
                     }else{
-                        ((CustomCardView) v).removeBorder();
-                        ((CustomViewEmpty) finalHolder.colorSub).removeBorder();
                         SelectedObserverService.getInstance().getIsSelected()[position] = false;
                     }
+                    updateBorderView();
                 }
             }
         });
